Swiat Heksagonalny wyswietlanie

+++ src/WirtualnySwiat/Punkt.java
@@ -3,26 +3,26 @@ package WirtualnySwiat;
 import WirtualnySwiat.Swiaty.Swiat;
 
 public class Punkt implements Cloneable{
-    private int x, y;
+    private double x, y;
 
-    public Punkt(int x, int y) {
+    public Punkt(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
@@ -71,6 +71,6 @@ public class Punkt implements Cloneable{
 
     @Override
     public String toString() {
-        return "(" + this.getX() + "; " + this.getY() + ")";
+        return "(" + (int) this.getX() + "; " + (int) this.getY() + ")";
     }
 }
